requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=1.1.0
pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
lxml>=4.9.0
tqdm>=4.65.0
tensorboard>=2.11.0
jiwer>=2.3.0
